APT/Applet: move applet managing into its own class
