+++ src/core/hle/applets/mii_selector.cpp
@@ -18,8 +18,8 @@ namespace HLE {
 namespace Applets {
 
 ResultCode MiiSelector::ReceiveParameter(const Service::APT::MessageParameter& parameter) {
-    if (parameter.signal != static_cast<u32>(Service::APT::SignalType::Request)) {
-        LOG_ERROR(Service_APT, "unsupported signal %u", parameter.signal);
+    if (parameter.signal != Service::APT::SignalType::Request) {
+        LOG_ERROR(Service_APT, "unsupported signal %u", static_cast<u32>(parameter.signal));
         UNIMPLEMENTED();
         // TODO(Subv): Find the right error code
         return ResultCode(-1);
@@ -43,13 +43,13 @@ ResultCode MiiSelector::ReceiveParameter(const Service::APT::MessageParameter& p
 
     // Send the response message with the newly created SharedMemory
     Service::APT::MessageParameter result;
-    result.signal = static_cast<u32>(Service::APT::SignalType::Response);
+    result.signal = Service::APT::SignalType::Response;
     result.buffer.clear();
-    result.destination_id = static_cast<u32>(Service::APT::AppletId::Application);
-    result.sender_id = static_cast<u32>(id);
+    result.destination_id = Service::APT::AppletId::Application;
+    result.sender_id = id;
     result.object = framebuffer_memory;
 
-    Service::APT::SendParameter(result);
+    SendParameter(result);
     return RESULT_SUCCESS;
 }
 
@@ -72,10 +72,10 @@ ResultCode MiiSelector::StartImpl(const Service::APT::AppletStartupParameter& pa
     Service::APT::MessageParameter message;
     message.buffer.resize(sizeof(MiiResult));
     std::memcpy(message.buffer.data(), &result, message.buffer.size());
-    message.signal = static_cast<u32>(Service::APT::SignalType::WakeupByExit);
-    message.destination_id = static_cast<u32>(Service::APT::AppletId::Application);
-    message.sender_id = static_cast<u32>(id);
-    Service::APT::SendParameter(message);
+    message.signal = Service::APT::SignalType::WakeupByExit;
+    message.destination_id = Service::APT::AppletId::Application;
+    message.sender_id = id;
+    SendParameter(message);
 
     is_running = false;
     return RESULT_SUCCESS;
