+++ src/core/hle/service/ns/ns.h
@@ -5,6 +5,7 @@
 #pragma once
 
 #include "core/hle/kernel/process.h"
+#include "core/hle/service/fs/archive.h"
 #include "core/hle/service/service.h"
 
 namespace Service {
